💄 UI: 핀 수정 모달창 따로 정의, 장소 말줄임 추가
